add left-right buttons to admin side
