+++ frontend/src/components/VolunteerDetailsPanel/index.css
@@ -1,16 +1,5 @@
 @import '../../styles.css';  /*Imports custom fonts*/
 
-.panel-button-icon {
-     padding: 0px;
-     background-color: transparent;
-     border: none;
-     cursor: pointer;
-}
-
-.panel-button-icon:hover {
-     filter:brightness(0.6);
-}
-
 .my-shifts {
      padding: 4px 12px;
      width: fit-content;
@@ -66,15 +55,6 @@
      height: 28px;
      border-radius: 4px;
      margin-right: 10px;
-}
-
-.button-icons {
-     display: flex;
-     flex-direction: row;
-     gap: 4px;
-     width: fit-content;
-     margin-left: auto;
-     margin-right: 36px;
 }
 
 /* Shift Details Panel */
