+++ frontend/src/components/VolunteerDetailsPanel/index.js
@@ -1,6 +1,4 @@
 import React from "react";
-import button_icon_next from "../../assets/images/button-icons/button-icon-next.png";
-import button_icon_prev from "../../assets/images/button-icons/button-icon-prev.png";
 import zoom_icon from "../../assets/zoom.png";
 import { formatImageUrl } from "../../api/imageService";
 import { COVERAGE_STATUSES } from "../../data/constants";
@@ -53,24 +51,6 @@ function VolunteerDetailsPanel({ classId, classList, setClassId, dynamicShiftBut
             })}
             </div>
         );
-    };
-
-    const handleToPrev = () => {
-        if (!classList || !classId) return;
-        const currentIndex = classList.findIndex((c) => c.class_id === classId);
-        if (currentIndex > 0) {
-            const prevClass = classList[currentIndex - 1];
-            setClassId(prevClass.class_id);
-        }
-    };
-
-    const handleToNext = () => {
-        if (!classList || !classId) return;
-        const currentIndex = classList.findIndex((c) => c.class_id === classId);
-        if (currentIndex < classList.length - 1) {
-            const nextClass = classList[currentIndex + 1];
-            setClassId(nextClass.class_id);
-        }
     };
 
     return (
@@ -163,22 +143,6 @@ function VolunteerDetailsPanel({ classId, classList, setClassId, dynamicShiftBut
                         {button.label}
                     </button>
                 ))}
-            </div>
-            <div className="button-icons">
-                <button className="panel-button-icon" onClick={handleToPrev}>
-                    <img
-                    alt="Previous"
-                    src={button_icon_prev}
-                    style={{ width: 16, height: 16 }}
-                    />
-                </button>
-                <button className="panel-button-icon" onClick={handleToNext}>
-                    <img
-                    alt="Next"
-                    src={button_icon_next}
-                    style={{ width: 16, height: 16 }}
-                    />
-                </button>
             </div>
         </div>
     )
